fix my_ingredients

+++ webapp/src/main/java/ar/edu/itba/paw/webapp/form/AddIngredientForm.java
@@ -21,7 +21,7 @@ public class AddIngredientForm {
         return amount;
     }
 
-    public void setAmount(int amount) {
+    public void setAmount(float amount) {
         this.amount = amount;
     }
 }
